code/requirements.txt
absl-py==2.1.0
aiohttp==3.9.5
aiosignal==1.3.1
annotated-types==0.6.0
anthropic==0.30.1
anyio==4.3.0
asttokens==2.4.1
attrs==23.2.0
backcall==0.2.0
beautifulsoup4==4.12.3
bert-score==0.3.13
bleach==6.1.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
comm==0.2.2
contourpy==1.2.1
cycler==0.12.1
datasets==2.20.0
debugpy==1.8.1
decorator==5.1.1
defusedxml==0.7.1
dill==0.3.8
distro==1.9.0
docopt==0.6.2
evaluate==0.4.2
executing==2.0.1
fastjsonschema==2.20.0
filelock==3.14.0
fonttools==4.53.0
frozenlist==1.4.1
fsspec==2024.5.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
huggingface-hub==0.23.3
idna==3.7
ImageHash==4.3.1
ipykernel==6.29.4
ipython==8.12.3
jedi==0.19.1
Jinja2==3.1.4
jiter==0.5.0
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
jupyter_client==8.6.1
jupyter_core==5.7.2
jupyterlab_pygments==0.3.0
kiwisolver==1.4.5
kornia==0.7.3
kornia_rs==0.1.5
lxml==5.2.2
MarkupSafe==2.1.5
matplotlib==3.9.0
matplotlib-inline==0.1.7
mistune==3.0.2
mpmath==1.3.0
multidict==6.0.5
multiprocess==0.70.16
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.3
nltk==3.8.1
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.5.40
nvidia-nvtx-cu12==12.1.105
ollama==0.2.1
openai==1.30.1
packaging==24.0
pandas==2.2.2
pandas-flavor==0.6.0
pandocfilters==1.5.1
parso==0.8.4
patsy==0.5.6
pexpect==4.9.0
pickleshare==0.7.5
pillow==10.3.0
pingouin==0.5.4
pipreqs==0.5.0
platformdirs==4.2.1
prompt-toolkit==3.0.43
psutil==5.9.8
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==16.1.0
pyarrow-hotfix==0.6
pydantic==2.7.1
pydantic_core==2.18.2
Pygments==2.18.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
PyWavelets==1.6.0
PyYAML==6.0.1
pyzmq==26.0.3
referencing==0.35.1
regex==2024.5.15
requests==2.32.3
rouge_score==0.1.2
rpds-py==0.20.0
safetensors==0.4.3
scikit-learn==1.5.1
scipy==1.13.1
seaborn==0.13.2
six==1.16.0
sniffio==1.3.1
soupsieve==2.5
stack-data==0.6.3
statsmodels==0.14.2
sympy==1.12.1
tabulate==0.9.0
threadpoolctl==3.5.0
tikzplotlib @ git+https://github.com/nicholasdehnen/tikzplotlib.git@188bfe0c505a89a4228817fdeba6e01ec2fec4dc
tinycss2==1.3.0
tokenizers==0.19.1
torch==2.3.1
tornado==6.4
tqdm==4.66.4
traitlets==5.14.3
transformers==4.41.2
triton==2.3.1
typing_extensions==4.11.0
tzdata==2024.1
urllib3==2.2.1
wcwidth==0.2.13
webcolors==1.13
webencodings==0.5.1
xarray==2024.6.0
xxhash==3.4.1
yarg==0.1.9
yarl==1.9.4
